+++ CHANGELOG.md
@@ -12,7 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add option to open files from outside OpenComic in a new window [`906a5b7`](https://github.com/ollm/OpenComic/commit/906a5b758685e0df2d67102856c1f3f74c53760c)
 - Add setting to use a custom folder for cache and temporary files [`5b7a3a1`](https://github.com/ollm/OpenComic/commit/5b7a3a150d65eef9ed92b4265c72818c30a29ec9)
 - AI image upscaling tool [`566295e`](https://github.com/ollm/OpenComic/commit/566295e86c3ecd0af95b022ddf7fbeab9ce7477d)
-- AI image descreen tool
+- AI image descreen tool [`aa88777`](https://github.com/ollm/OpenComic/commit/aa88777d5a57c1e8819bed8284d844b3b05fdaf1)
+- AI image artifact removal tool
 
 ##### 🐛 Bug Fixes